parseError вынесен в common, внесены изменения в редьюсер user
